==> notebook/1_data_processing.ipynb <==
# Prediabetes Risk Assessment
---

### Introduction
---
This project aims to assess the risk of prediabetes among Indian youngsters aged from 15-25 based on various health and lifestyle factors. The dataset contains information on age, gender, region, family income, family history of diabetes, genetic risk score, BMI, physical activity level, dietary habits, fast food intake, smoking status, alcohol consumption, fasting blood sugar, HbA1c, cholesterol level, sleep hours, stress level, and screen time.

By analyzing this data, I aim to identify key factors that contribute to the risk of prediabetes and develop predictive models to assess individual risk levels. This can help in early detection and preventive measures to reduce the incidence of diabetes.

---
## 1.) Import Required Packages

####  Importing Pandas, Numpy, Matplotlib, Seaborn and Warings Library.

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt 
import warnings
import os
warnings.filterwarnings('ignore')

---
## 2.) Data Collection
- Dataset Source -https://www.kaggle.com/datasets/ankushpanday1/diabetes-in-youth-vs-adult-in-india
- The data consists of 22 column and 100000 rows.

#### Import the CSV Data as Pandas DataFrame

df=pd.read_csv('../data/raw_data/indian_youngsters_health_data.csv')

#### Show Top 5 Records

df.head()

#### Shape of the dataset

df.shape

### 2.1 Dataset information

- ID: Unique identifier for each record.
- Age: Age of the individual (15-25 years).
- Gender: Gender of the individual: Male, Female, or Other.
- Region: Region in India where the individual resides.
- Family_Income: Annual family income in INR.
- Family History of Heart Disease: Indicates if there is a family history of diabetes: Yes or No.
- Parent_Diabetes_Type: Indicates if parents have diabetes and its type: Type 1, Type 2, or None.
- Genetic_Risk_Score: A numeric score (1-10) indicating genetic predisposition to diabetes.
- BMI: Body Mass Index, a measure of body fat based on height and weight.
- Physical_Activity_Level: Level of physical activity: Sedentary, Moderate, or Active.
- Dietary_Habits: Dietary habits: Healthy, Unhealthy, or Mixed.
- Fast_Food_Intake: Frequency of fast food intake: Rarely, Occasionally, or Frequently.
- Smoking: Smoking status: Yes or No.
- Alcohol_Consumption: consumption of alcohol: Yes or No.
- Fasting_Blood_Sugar: Fasting blood sugar level in mg/dL.
- HbA1c: Glycated hemoglobin level in %.
- Cholesterol_Level: Cholesterol level in mg/dL.
- Prediabetes: Indicates if the individual has prediabetes: Yes or No.
- Diabetes_Type: Indicates the type of diabetes: Type 1, Type 2, or None.
- Sleep_Hours: Average hours of sleep per night.
- Stress_Level: A numeric score (1-10) indicating Level of stress.
- Screen_Time: Average hours of screen time per day.

---
## 3.) Data Checks to perform

- Check Missing values
- Check Duplicates
- Check data type
- Check the number of unique values of each column
- Check statistics of data set

### 3.1 Check Missing values

df.isna().sum()

### Missing Values are Found So Handling The Missing Values

df['Parent_Diabetes_Type'].fillna('No Diabetes',inplace=True)
df['Diabetes_Type'].fillna('No Diabetes',inplace=True)

### Ensure No More Missing values

df.isna().sum()

#### There are no missing values in the data set

### 3.2 Check Duplicates

df.duplicated().sum()

#### There are no duplicates  values in the data set

### 3.3 Check data types

# Check Null and Dtypes
df.info()

### 3.4 Checking the number of unique values of each column

df.nunique()

### 3.5 Check statistics of data set

df.describe().T

df.describe(include='object').T

df.columns

---
## 4.) Saving the Processed data

save_location = '../data/processed_data/indian_youngsters_health_data.csv'
os.makedirs(os.path.dirname(save_location), exist_ok=True)  # Ensure the directory exists
df.to_csv(save_location, index=False)
